requirements-dev.txt
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
black
codecov
coverage
fastapi[all]
flake8
isort
pytest
pytest-xprocess
py
sphinx
ipython
numpydoc
sphinx
sphinx_rtd_theme
requests
